split tasks.py into package
